Organize classes within the hex.genmodel.algos package
